Add roi.project and roi.transform packages
