+++ src/main/java/net/imagej/omero/rois/project/OMEROZTCProjectedRealMask.java
@@ -23,7 +23,7 @@
  * #L%
  */
 
-package net.imagej.omero.rois;
+package net.imagej.omero.rois.project;
 
 import net.imglib2.AbstractEuclideanSpace;
 import net.imglib2.RealLocalizable;
