+++ src/main/java/net/imagej/omero/rois/transform/RealTransformUnaryCompositeRealMaskRealIntervalToShapeData.java
@@ -23,11 +23,12 @@
  * #L%
  */
 
-package net.imagej.omero.rois;
+package net.imagej.omero.rois.transform;
 
 import java.lang.reflect.Type;
 import java.util.function.Predicate;
 
+import net.imagej.omero.rois.RoiConverters;
 import net.imglib2.realtransform.AffineTransform2D;
 import net.imglib2.roi.Operators.RealTransformMaskOperator;
 import net.imglib2.roi.composite.CompositeMaskPredicate;
