+++ src/test/java/net/imagej/omero/rois/DataNodeToROIDataConversionTest.java
@@ -40,7 +40,13 @@ import net.imagej.axis.TypedAxis;
 import net.imagej.omero.OMEROLocation;
 import net.imagej.omero.OMEROService;
 import net.imagej.omero.OMEROSession;
+import net.imagej.omero.rois.project.OMEROZTCProjectedRealMaskRealInterval;
+import net.imagej.omero.rois.project.OMEROZTCProjectedRealMaskToShapeData;
+import net.imagej.omero.rois.project.ProjectedOMERORealMaskToShapeData;
 import net.imagej.omero.rois.rectangle.ClosedOMERORectangle;
+import net.imagej.omero.rois.transform.RealTransformUnaryCompositeRealMaskRealIntervalToShapeData;
+import net.imagej.omero.rois.transform.TransformedOMERORealMaskRealInterval;
+import net.imagej.omero.rois.transform.TransformedOMERORealMaskRealIntervalToShapeData;
 import net.imagej.space.DefaultTypedSpace;
 import net.imagej.space.TypedSpace;
 import net.imagej.table.Table;
